I have added a new Unity Project
All I did was add some folders and put in a scene that is set up in a 2D
environment. So nothing really I just needed something to save in the
assets so that the folder will open as a unity project for y'all when
y'all go to open it.